Update synth_recipes.sql
For each change, standards were: Desynth with FFXIAH listing the skill level as 255, BGWiki having a specific level, recipe and results match.

Recipes with a specific BGWiki skill level but a mismatch in yield or recepie had their skill level set, but remained commented out and had comment added at the end of the line.