Update synth_recipes.sql
For this round, I used the following criteria:

1: Item has a specific desynth recipe and yield listed on its BGWiki page (wihtout a level) AND Item has a synthesis recepie
2: Item was previously updated in level but not activated due to yield mismatch.

In cases where the item is craftable, but the desynth recepie had no level assigned to it, the level was set to match the synthesis level, and a comment was added in case future evidence proves that the level should be different.

In cases where the yield shown on FFXIAH/the current database were similar, but not exact matches, (More HQ result information on the BGWiki page primarily), the BGWiki data was used. This includes previously identified mismatched yields, and they have been activated as well.